requirements.txt
ansi2html==1.8.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
findspark==2.0.1
Flask==2.2.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
plotly==5.15.0
py4j==0.10.9.7
pyarrow==12.0.1
pyspark==3.4.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
venv-pack==0.2.0
Werkzeug==2.2.3
